Bug 528251 - view CTabFolders result in not highlighted part stack tab
When a view contains a selected CTabFolder, selecting another view in
the same part stack results in non-highlighted view tab. This is the
case, since CTabFolder.onActivate/onDeactivate will tell ancestor
CTabFolders to highlight resp. not highlight. Selecting another view
results in deactivating the nested CTabFolder and so removing
highlighting from the part stack tabs. This regression was introduced
with the fix for Bug 474444.

With this change, highlighting is toggled only for the CTabFolder which
is being activated/deactivated. The change was verified both on Windows
and on RHEL+GTK3.14. The problem from Bug 474444 is not observed.

Change-Id: I6fd4bf7f049e84d95bd5e83332a6484164623576
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>