Bug 525763: [GTK3] [dark theme] Vertical bar to the left of line numbers is not dark
This is a clean revert of certain aspects of bug 477950.The root issue
of the bug is caused by the Canvas in AnnotatedChangeColumn getting its
background color from the wrong parent widget. Also added a background
state check in certain widgets that manage their backgrounds separately
(i.e. outside of Control).

Change-Id: Id5b12d57ff3073e86eceb6c5736fbbea93551a30
Signed-off-by: Eric Williams <ericwill@redhat.com>