Bug 528251 - Test to check tab highlight with nested CTabFolders
When a view contains a selected CTabFolder, selecting another view in
the same part stack results in non-highlighted view tab. This is the
case, since CTabFolder.onActivate/onDeactivate will tell ancestor
CTabFolders to highlight resp. not highlight. Selecting another view
results in deactivating the nested CTabFolder and so removing
highlighting from the part stack tabs.

This is a test for the missing highlight.


Change-Id: Iad92b122883068114ddeebe96bc422da206d2c85
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>