Bug 528415 - CTabFolder tabs vanish after mouse-over
If a CTabFolder specifies SWT.RIGHT_TO_LEFT as one of its styles,
mouse-over on a tab will make the tab not draw. This painting artifact
remains until another window becomes active. See the bug snippet.

This is due to wrong clipping set to the draw event in
Control.gtk_draw(). The event bounds are correctly set in the case of
SWT.MIRRORED (derived from SWT.RIGHT_TO_LEFT), but for the clipping area
SWT.MIRRORED is not taken into account. This is a regression coming from
bug 421127.

This change ensures the event clipping equals the event bounds, on draw
event creation. This ensures the tabs of the CTabFolder are redrawn.

Change-Id: I3bb9049bac98c5f5d360135771ef43ef1cc1a9b1
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>