Bug 483791: [GTK3] setBackground overrides GC drawing in PaintListener
This bug happens on GTK 3.9+, due to the fact that the GTK drawing
events are connected to the wrong widget. Changing the background color
on those widgets does not trigger the proper re-drawing. As far as I can
tell this has been a bug for awhile and is not a regression. Only Button
and Label are affected.

Solution: connect the paint signal/GTK drawing events to
the appropriate handles in Button and Label.

Change-Id: Ib6b513493cbb74680facd6ebc81569c3fa066379
Signed-off-by: Eric Williams <ericwill@redhat.com>