Fix Library.java to stick to ASCII symbols for builders.
Build broke with:
/opt/public/hipp/homes/genie.platform/workspace/tmp/check.compile.master/build/org/eclipse/swt/internal/Library.java:367:
error: unmappable character for encoding ASCII
 * Dev note: (17??12??07) This has been developed and throughly tested
on GTK. Designed to work on Cocoa/Win as well, but not tested.
                ^
/opt/public/hipp/homes/genie.platform/workspace/tmp/check.compile.master/build/org/eclipse/swt/internal/Library.java:367:
error: unmappable character for encoding ASCII
 * Dev note: (17??12??07) This has been developed and throughly tested
on GTK. Designed to work on Cocoa/Win as well, but not tested.
                 ^
/opt/public/hipp/homes/genie.platform/workspace/tmp/check.compile.master/build/org/eclipse/swt/internal/Library.java:367:
error: unmappable character for encoding ASCII
 * Dev note: (17??12??07) This has been developed and throughly tested
on GTK. Designed to work on Cocoa/Win as well, but not tested.
                    ^
/opt/public/hipp/homes/genie.platform/workspace/tmp/check.compile.master/build/org/eclipse/swt/internal/Library.java:367:
error: unmappable character for encoding ASCII
 * Dev note: (17??12??07) This has been developed and throughly tested
on GTK. Designed to work on Cocoa/Win as well, but not tested.
                     ^

Change-Id: I431151233a8ec853fccc75636b53f452cf03b760
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>