Bug 528155: [GTK] Table/Tree forgets to set SWT.SELECTED bit on PaintItem
Modifying the SWT.SELECTED bit in an EraseItem listener causes the
SWT.SELECTED bit to not be sent to the PaintListener. This affects
PaintListeners that draw based on Table/Tree selection state.

The fix is to cache the SWT.SELECTED bit using the existing boolean
"wasSelected", and set the SWT.SELECTED later on before firing the
PaintEvent.

Change-Id: I161452695b3041a37a66eca5e3de75035a56f9a7
Signed-off-by: Eric Williams <ericwill@redhat.com>