Bug 529227: FileDialog setFilterExtensions() doesn’t allow to choose files that were filtered with two dots in their extension
In Cocoa, only the part of the file name after the last extension
divider (.) is considered as extension. NSString.pathExtension() in Cocoa too
returns only this last part. But, SWT FileDialog supports
extensions with more than one (.). When files with extensions which have
more than 1 separator are filtered, they are not shown as enabled in the
File Open Dialog.
The workaround is to add only the extension after the last (.) as the allowed
file type.

Change-Id: I578fa557ffaff59113d22d3784f7ee1825599a4f