Bug 513761: [GTK3] Tree with columns and with CHECK and VIRTUAL styles renders checkboxes in 2 columns
Part of the fixes for bug 480261 involve re-creating Tree renderers when
the SWT.VIRTUAL style is specified. This fix lead to a regression where
checkboxes are rendered in the columns other than the first.

This patch verifies that checkbox renderers are re-created only for the
first column in the Tree, and does not break the fixes for bug 480261.

Change-Id: Ia3ee77a476684e4ebad264387690505a2820aeb4
Signed-off-by: Eric Williams <ericwill@redhat.com>