Bug 519295: [GTK3] Invisible settings button in validation preferences
Instead of setting the image dimensions for each column, set them in the
parent and use those dimensions for every column. Every image renderer
should be the same size anyways, regardless of column -- the size is
determined by the dimensions of the first image set.

The bug was caused by parent.pixbufSizeSet being true, but the column
renderers being size 0x0. The renderers in the third column never got
sized because the column renderer size was never set. Funnily enough the
ability to check for this was added in Tree but never implemented. Both
Table and Tree have been fixed now.

Change-Id: Ibad077cb4de21071bb22b75446881d63da0590e0
Signed-off-by: Eric Williams <ericwill@redhat.com>