Bug 481485: [GTK3.10+] Paint listener for button/label does not work
On GTK3.10+, Buttons and Labels do not receive the proper draw signals.
This is due to them drawing on their SwtFixed parents (fixedHandle), as
GtkButton and GtkLabel do not have their own GdkWindows.

This patch reverts parts of bug 483791 and fixes this bug by overriding
windowProc() in Button and Label. There we check if the incoming handle
is the paintHandle, and draw accordingly.

Bug snippet is attached in this patch.

Change-Id: I7c415985248ca5fda9e38aaf9b0c08ae834d762d
Signed-off-by: Eric Williams <ericwill@redhat.com>