Bug 530969: [GTK3] Control.print() doesn't seem to work
In GTK3, gtk_widget_draw() relies on the priv->alloc_needed flag to be
set to TRUE. This is an internal flag in a private struct so we cannot
access it. The fix is to the get the topHandle's allocation, and simply
re-set it. This triggers an allocation change without actually changing
the size of the widget.

Tested on GTK3.8+. No JUnit test failures.

Change-Id: I81df53bfceaa988a2cce574ef7629c64c0ec6fb9
Signed-off-by: Eric Williams <ericwill@redhat.com>