Bug 511133: [GTK3.10+] Some critical icons missing on Ubuntu 16.04
This patch fixes Table/Tree drawing in EGit. It patch reverts the
changes introduced for bug 438505, which causes this regression. This
patch also contains fixes to prevent bug 438505 from being
re-introduced.

EGit drawing was broken because TableItem.getBounds() was taking the
position of the header into account. This is because we were drawing on
the fixedHandle instead of the GtkTreeView. The positioning of the Table
header broke simple drawing cases like Snippet051TableCenteredImage
(JFace snippet). Reverting the fix back to drawing on the handle has
fixed that issue.

Unfortunately this means bug 438505 is re-introduced. To fix that issue
I have tweaked the re-parenting logic found in
Table/Tree.setParentWindow(). GTK3.10+ onward only draws on toplevel or
native GdkWindows. This means the Text widgets used for Table/Tree
editing continue to get events, but they are not drawn. The solution to
this problem is call gdk_window_raise/lower on these Text widgets when
setVisible() is called. This raises and lowers the Text's GdkWindow when
the widget's visibility is changed, ensuring that it actually is drawn
when visible, and not when hidden. This isn't a huge change as the
GdkWindow re-parenting already existed in Table/Tree.setParentWindow(),
this patch just builds on that.

I have tested against the following cases/snippets:
Snippet88
Snippet019 (JFace snippets)
Snippet025 (JFace snippets)
Snippet026 (JFace snippets)
Snippet051 (JFace snippets)
use case from bug 438505
use case from bug 436324
use case from bug 460581
use case from bug 458630

Additionally, using a child Eclipse I have verified that the missing
EGit icons have returned. No AllNonBrowser JUnit tests fail.

Change-Id: I02b3fb30037a8f0f74de79144f8e563f5284d571
Signed-off-by: Eric Williams <ericwill@redhat.com>