Bug 528498 - Combo menu is narrower than combo
In order to fix Bug 438992, wrap width of 1 was set for read-only
combos. This results in drop-down menus which are narrower than the
combo itself. Seen e.g. in the Eclipse Help preference page, where the
combos are as long as the page, but their entries are relatively short.
The drop-down menu is then only as short as the longest entry.

This change ensures that the GTK style option
-GtkComboBox-appears-as-list is used, for GTK versions above 3.22.5.
This changes the drop-down menu style, preventing both Bug 438992 and
Bug 528498.

The option cannot be used in earlier versions, since due to a bug, the
drop-down menu cannot be scrolled.

Change-Id: I96aecae46c6edef65e69a9ae590e713bf607dd4b
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>