Bug 378202 - Two-finger tap to simulate right-click results in multiple click events
Control.menuForEvent needs to return the menu to be shown for the right
click instead of 0, or the right click will be propagated up the chain.
If the menu is result of getMenu(), it is returned.
If it is created in the SWT.MenuDetect event listener, it is not shown
yet, but registered in display.popups. When this is detected it is
removed from that array and returned.

Change-Id: Ia2bf133f61b0ce74d036c9188fbf1dd36ca76508
Signed-off-by: Till Brychcy <register.eclipse@brychcy.de>