Bug 531593 – [Webkit2] Crash in webkit on hover
I can't reproduce the reported issue, but I can see a rare corner
race condition where such a crash could occur.

I.e, the race condition could be that webkit has internally
disposed itself but we might be referencing it accidentally
still.
The fix is to guard against this. The fixes don't break anything
as the jvm crash is occurring on webkit's disposal anyway.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=531593
Change-Id: Ia7394ce301b237dcf5ac43edb6590f0f5003e49f
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>