Bug 431423: [GTK3] Context menu should appear by pointer when invoked via mouse
Some context Menus on GTK3.22+ had an issue where the content of the
menu was incorrect: often the content of the previous menu would be
shown, or some menu items were missing entirely.

This fix ensures the proper event process is followed so that
dynamically populated menus have enough time to be created before being
shown. This is done by reverting a fix which removed
display.[add/remove]Popup logic in Menu.setVisible().

Change-Id: Ice2fe3a1b8ecb58035dbe69e5de1bf9749fc22c2
Signed-off-by: Eric Williams <ericwill@redhat.com>