Bug 431423: [GTK3] Context menu should appear by pointer when invoked via mouse
Some context menus are positioned incorrectly when the size (number of
items) of the menu changes between clicks.

The cause of this bug is due to the way SWT populates its menus: often
it is done asynchronously outside of SWT based on an SWT.Show listener.
The result is that the menu items are added/removed just before showing,
and GTK sometimes doesn't have enough time to properly adjust the size
of the menu internally. Specifically, the GdkWindow of the toplevel
widget (GtkWindow) for the menu lags in height. Since GTK thinks the
menu is taller/shorter than it is, the menu is positioned at the wrong
y-coordinate.

The fix is to check if the number of menu items has changed from one
pop-up to the next. If so, calculate the preferred height of the menu
and resize the GdkWindow of the menu's toplevel GtkWindow to reflect
this change. This way the internal sizing calculations in GTK will be
accurate and the menu will be positioned correctly.

The bug is fixed on GTK3.22 with X11 only, as this uses the new GTK3.22
GtkMenu API. Older GTK3 versions use the outdated GtkMenu API and fixing
these cases would not be worth the effort. Wayland is not affected by
this bug.

Reliable steps to reproduce the issue can be found here:
https://bugs.eclipse.org/bugs/show_bug.cgi?id=431423#c19

Tested JFace/SWT menu snippets, AllNonBrowser JUnit tests, and
ControlExample on GTK3.22 with X11.

Change-Id: I392983438cc32b9455c53cc464626294f3ec72f9
Signed-off-by: Eric Williams <ericwill@redhat.com>