Bug 531667 - [GTK3] Cannot draw Canvas with Control.print(GC)
Calling Canvas.pring(GC) results in no painted canvas on the GC. This is
caused by a regression from the fix for Bug 421127. Control.print
propagates the GC.handle as the last argument to Control.gtk_draw. With
changes for Bug 421127 in Control.gtk_draw, the handle is no longer
propagated to the actual onPaint event. The canvas (and canvas derived
widgets) therefore no longer paint themselves on the GC.

The problem which was fixed with the change in Control.gtk_draw, for Bug
421127, was that CTabFolder tabs were not painted. This is the case,
since GC.getClipping returned broken values with GTK 3.10 and above.
CTabFolder checks for intersection with the clipping region and doesn't
paint itself.

The clipping region was not computed properly, due to transformations
done with the device space transformation matrix provided by Cairo. From
GTK 3.10 onward, this matrix contains absolute coordinates of the
widget, with respect to the shell. This causes the method to
malfunction.

With the fix for Bug 421127, the Cairo matrix is no longer propagated.
As a result GC.getClipping only works with the identity matrix,
resulting in correct clipping.

This change reverts the change for Bug 421127 in Control.gtk_draw, for
GTK versions 3.14 and above. The GC.getClipping method is adapted to
first transform the local GC bounds to device space, before computing
the intersection with the clipping region. This is done for GTK 3.14 and
above.

We validated the change with GTK 3.14 and GTK 3.22. A bug snippet is
provided, show-casing clipping region and Cairo device space
transformation matrix.

Change-Id: Ib1d6bb866bcffa010a9e8f2a5e4296eba4e1d852
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>