Bug 529431: [GTK3.10+] Snippet294 fails to draw Region
Draw the region on the Control using cairo_clip(). This doesn't use the
GdkWindow API which no longer works on GTK3.10+.

The only caveat is that input handling is not blocked. For example:
setting the region on a button will not prevent click events within the
region from registering. This needs to be investigated and fixed.

Tested on GTK3.8 - GTK3.22. IDE looks fine, no AllNonBrowser JUnit tests
fail.

Change-Id: If6f6bdc1d349cfb3ab42be3977e5c608e29df43e
Signed-off-by: Eric Williams <ericwill@redhat.com>