Bug 530317: [GTK3.20+] Java formatter editor has cut-off buttons
This bug is caused by a corner case where the minimum size is larger
than the natural size. In these cases the widget will end up too large,
as only the minimum size is taken into account.

When performing sizing calculations for GTK3.20+, use the smallest of
the minimum and natural sizes to prevent oversized widgets.

Tested on GTK3.20+ using Eclipse and the AllNonBrowser JUnit tests. No
additional allocation warnings are printed in the console.

Change-Id: I30a7ef75abb135b1a23402a834824241b6eb055f
Signed-off-by: Eric Williams <ericwill@redhat.com>