Bug 500703: [GTK3.20+] Combo with SWT.READ_ONLY is garbled upon re-size
In GTK3.20+ two things inside GTK3 changed. First, certain widgets have
their clips unioned with the clips of their parent widget. This can
cause overdrawing in widgets like Combo. Second, the internal widget
hierarchy of Combo changed, and most of the resizing is done by a GtkBox
instead of GtkComboBox itself.

These two changes combined caused this bug. The overdrawing causes the
Combo to draw over its neighbouring widgets (like a label). The resize
issue means that resizing the widget doesn't cause it get smaller: the x
coordinates of the allocation and clip just become negative and the
whole widget shifts to the left.

The fix is to adjust the clip and allocation of the Combo and its
children from the parent Composite's draw handler. This ensures that the
Combo is drawn with correct coordinates and that it actually shrinks
when resized, instead of just shifting.

Tested on GTK3.22, X11 and Wayland. Window -> Preferences -> General ->
Keys behaves as expected, and no AllNonBrowser JUnit tests fail.

Change-Id: Ic603acfda924b0ca36b4f0ba07048f6762c99270
Signed-off-by: Eric Williams <ericwill@redhat.com>