[Tests] Use dedicated SWT native lib cache folder
SWT tests often also involve testing a new snapshot of the native
libraries. As SWT usually puts the libraries in a cache under
user.home, it makes SWT tests more likely to used the cached version
then the one that inside platform-specific fragments.
SWT Tests are now configured to use a different user.home to not interact
with the default cache directory.

Signed-off-by: Mickael Istria <mistria@redhat.com>