Bug 529431: [GTK3.10+] Snippet294 fails to draw Region
This patch fixes drawing issues when setRegion is called on widgets that
are children of Scrollable (like List, for example).

The fix is to check if there are incoming draw events for paintHandle()
in Control. If this is the case, and setRegion has been defined, we call
gtk_draw(). This ensures that the clipping is applied to the right
handle and that the region is drawn.

Tested on GTK3.22. Compared behaviour with GTK3.8: most widgets work
with the exception of Table/Tree and a few others. No AllNonBrowser
JUnit tests fail. Functionality for Shell.setRegion() remains
functional.

Change-Id: Ife9a1688e58eef3c98cc6f74ff7d1bac1eac9986
Signed-off-by: Eric Williams <ericwill@redhat.com>