Bug 522140: [GTK3] Horizontal SashForm resize problem
Use lastX instead of rect.x in Sash.gtk_motion_notify_event(). LastX is
set to rect.x only when event.doit is true, which is the case in 99% of
use cases. However when resizing quickly, event.doit is not true, so
calling setBounds() on rect.x is incorrect.

Using lastX checks for event.doit which avoids this bug by not calling
setBounds() with rect.x as the x coordinate.

Tested on GTK3.22 in the IDE and ControlExample. No AllNonBrowser JUnit
tests fail.

Change-Id: I4661cae8df947a953e273b48d851254579fd163b
Signed-off-by: Eric Williams <ericwill@redhat.com>