Bug 533799 - shrinking trees/tables can cause errors on standard error
Resizing a GTK tree view so that the column header and the horizontal
scrollbar overlap causes pixman bug error messages on standard error.
Using standard GTK3 API, it doesn't seem possible to resize the table
the tree view in such a way. SWT on the other hand allows this.

This change ensures that the SWT fixed which contains the tree or table
is resized as requested by resize operation, while the tree or table
retains some minimum height to avoid the error. In result, parts of the
horizontal scrollbar are no longer visible, when the tree or table is
shrinked below about 40px, the column headers and the horizontal
scrollbar are visible.

Change-Id: I3b39596a0f4eba3e1bbb259ab30d9d8744bb62c1
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>