Bug 533833 – [Webkit2] Should not use G_TYPE_CHECK_INSTANCE_TYPE(..)
I adopted the code so that webkit2 no longer uses the macro.
But I didn't take the time to remove it's use from webkit1
as I don't think it's worth the time.

I left some notes warning not to use the macro.
Upon removal of webkit1, it should be removed.

Tests:
- Fedora 27, Webkit2.18, jUnits & child eclipse.
- Not tested on webkit1, but should work as there are no logic changes.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=533833
Change-Id: I825825c02413a35592dcf3e21114ed774cf146f0
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>