Bug 534204 - [GTK3] Scale with GridLayout height hint has no line
An SWT Scale for which a low height hint is specified via GridData does
not show its line on GTK 3.22. It only shows its trough. This is also
observed when the parent of the Scale has a GridLayout and has a small
enough size (e.g. due to resizing).

This change ensures the scale is never reduced to below 26px. Resizing
of the scale is not impacted, since the GTK scale resides in an SWT
fixed which can still be resized at will. When resized below 26px, parts
of the scale start to become not visible.

Change-Id: I5d0b18783b5f126ca2fee6918b885480cbcb977e
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>