Bug 534007: [GTK3] Support for detecting themes like in 4.7.3 (but without the memory leaks)
Check for the presence of the GTK_THEME environment variable. When
loading system colors from the GTK CSS theme, we can make use of this
information to load the proper theme information and dark variants, if
necessary.

Tested with the Adwaita and Arc themes (and their dark variants) on
GTK3.22, X11 and Wayland. No AllNonBrowser JUnit tests fail.

Change-Id: If3697495a941ac4e26fb2d714d4b221da65325a1
Signed-off-by: Eric Williams <ericwill@redhat.com>