Bug 515773 – [wayland] Text widget not able to grab focus with SWT.ON_TOP shell (fixing crash of earlier patch).
At the time of the original patch, gdkSeat grab only worked on hidden
gdkWindows which were shown with handler, so Ian made gtkWindows
invisible before grabbing focus.

Having invisible windows leads them to be unrealized and
an unrealized window is passed to wayland, which crashes due to a
a segfault.

In newer gtk3 (3.22) this is no longer the case, seat grabbing works
on visible windows, so not hiding windows fixes the gnome crash
and doesn't break things.

Also I'm making the callback static instead of instance, because
otherwise there is a memory leak every time you raise the shell.

Tests:
- Child eclipse works well, including javadoc and outline. No crashes.
- Original bug snippet works well.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=515773
Change-Id: Ief9e066bbd488413ffbbe752f2b2c9ce561f532f
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>