Bug 534555 - [GTK] Browser stops responding and throws an error on dispose (Browser crash)
I found an subtle issue in javascript execution logic.

The javascript execution logic is canceled if browser's *shell* is
disposed.

This is wrong, the javascript logic should check if the *browser*
is disposed, not the browser's shell because you can have the
situation where the shell is not disposed but the browser is.

Also updating error message to clarify slow javascript execution for
users.

I can't reproduce the issue locally, so I can only assume this fixes
the reported issue.
But I tested jUnits and child Eclipse and didn't have issues.

Change-Id: Icae4ec6c28dd17fb1e555f24fbd5b0b984702917
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>