Bug 534692 - Scrollbar area in compare editor too small
The fix for bug 533799 causes a regression, which manifests as a
too-short vertical scroll bar in the Compare Editor.

The regression is caused by changing the implementation of
Scrollable.hScrollBarWidth() to return the width instead of the height
of the computed GTK preferred size requisition. This was due to an
oversight during refactoring.

This change restores the original API functionality, in return fixing
the layout of the compare editor.

Change-Id: I776df3efd8811b8a5fd2360691396d72cc3a6a3f
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>