Bug 498217 [GTK3][DnD] Dragging parts does not show rectangle (fix)
As of Gtk 3.9.1, Commit a60ccd3672467efb454b121993febc36f33cbc79,
off-screen GDK windows are not processed.
Because of this gtk doesn't send move events to SWT. Platform.UI
uses an off-screen tracker for SWT.MOVE events to draw custom
rectangles on a transparent shell for it's part-drag-preview.

Drawing/updates for these broke because tracker is off screen
and no longer sent mouse move events.

Solution: If a tracker is to move off-screen, then instead draw it 1x1
and make it transparent.

Tests:
- Child eclipse. Now part-preview works when dragging editor/views
around.
- Snippet 31/23 (tracker) work as before.

Note, this fix is only for X11. On Wayland part-preview is still broken
due to bug 535083. This bugfix seems to make part-preview
more fluent on wayland, but there's still a black shell over eclipse.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=498217
Change-Id: Ia69c494d3d08b9565e6647073bbfd4870f20ade3
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>