Bug 535064 - [HiDPI][GTK3] Device#getDPI method returns an incorrect value on x2 scaled screen
- fixed regression reported by Bug 535977 and Bug 535992

The earlier implementation used till 4.8M4 used to return 96 in all cases.
Here is the documentaion priovicded for gdk_screen_width_mm ()
"Returns the width of the default screen in millimeters. Note that on many X servers this value will not be correct."

Now around M4 we migrated SWT builds to foundation at that time the OS and GTK versions have been upgraded.
This caused gdk_screen_width_mm () start returning 0.

The new API is available from gtk 3.22 onwards so I am using new api from 3.22. but for previous versions of 
GTK3 we are returning 96 as it used to be previously. 


Change-Id: I7f7a1af6f3955e74430ac5a27146cecbf36cd56f
Signed-off-by: Sravan Kumar Lakkimsetti <sravankumarl@in.ibm.com>