Bug 475784: half transparent composite as overlay composite not work with gtk3
We have to handle some corner case drawing situations, since
Control.setRegion() now uses Cairo entirely. In particular, we need to
handle the case where a user is using an SWT.NO_BACKGROUND Composite as
an overlay. This means the Composite has a region set, and a
paintListener attached to perform custom drawing.

Previously, Control.setRegion() "physically" manipulate the Composite's
GdkWindow to prevent any drawing outside the given region. Since Cairo
is now responsible for drawing the set region we have to handle this
geometry ourselves.

This patch fixes numerous issues all rolled into one. First, it adds a
field in GCData which corresponds to the Cairo region. This is set in
gtk_draw() so that GC.fill* methods will clip themselves accordingly and
not fill where they aren't supposed to.

Second, it checks for a PaintListener, NO_BACKGROUND, and a region being
set. If these conditions are met, we draw a transparent background in
Control.drawBackground() to ensure the region and custom drawing
succeeds.

Last, we copy the region at the Cairo level to ensure there are no
crashes when calling setRegion(region) followed by region.dispose().

Tested on GTK3.22 using the snippet attached and a child Eclipse. No
AllNonBrowser JUnit tests fail.

Change-Id: Icdd04d8a4e0e504a5ee69688ba304af3956e4f77
Signed-off-by: Eric Williams <ericwill@redhat.com>