Bug 531341: [Webkit2] Can't access self signed web sites using internal web browser on fedora 27
Added functionality that prompts a user if they'd like to ignore TLS
errors for certain hosts. If a user visits a website that uses a
self-signed certificate, they will be prompted to either ignore TLS
errors for that host, or to go back.

Tested on GTK3.22 with Webkit2.20. No AllBrowser JUnit tests fail, and
Snippet128 works as expected.

Change-Id: I7c7a07ae27017a70e4a06c720f5640d094a6114c
Signed-off-by: Eric Williams <ericwill@redhat.com>(cherry picked from commit a8f4d8012ebe238243cfee7b3a063bd66a2b6f00)
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.swt/+/187113
Tested-by: Niraj Modi <niraj.modi@in.ibm.com>
Reviewed-by: Niraj Modi <niraj.modi@in.ibm.com>