Bug 395296: [GTK] Incorrect ToolBar width on Linux
Set the "homogeneous" property to false by default for all ToolItems.
This prevents a large ToolItem from dictating the size of all other
ToolItems belonging to the same ToolBar. For some reason, this property
is enabled by default in GTK3.

Tested using a child Eclipse, the Git rebase interactive view, and the
snippets attached to this Gerrit. No AllNonBrowser JUnit failures are
observed.

Environment: Fedora 30, GTK3.24.

Change-Id: I437309217fb5e706bce04e82a63a227e1a2baef0
Signed-off-by: Eric Williams <ericwill@redhat.com>