Bug 536912 - Optimize Display.initializeSystemColors
Initialize the theme provider and convert it to css only once in the
initializeSystemColors method instead of repeating it multiple times.
Pass the css string directly to methods that need it instead of
refetching it from theme provider again.
This required some tweaks to Control to pass the CSS directly to
gtk_css_parse_[background/foreground] instead of the theme provider.
With the above changes the time taken to do "new Display()" is reduced
from 950 to 270 ms.

Change-Id: Ie0996473cab2cc3b0c6d60084043073486b348ed
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>