Bug 536900 - [GTK3] New java class dialog renders a strange gray box as separator
For GTK3.20+, GtkSeparator with GTK_ALIGN_FILL renders a gray box that
takes the entire area rather than a single line. The fix is to set the
valign and halign to GTK_ALIGN_CENTER for the horizontal and vertical
separators, respectively, after creating the handles.

Tested with New Java class Dialog and Snippet37.

Change-Id: Ia11e917885af6cce61a2b21961f4bd27c4cbb910
Signed-off-by: Xi Yan <xixiyan@redhat.com>