Bug 536934 - [GTK3] Tree does not allow painting the full item
Changes made to custom drawn trees which updates event.x from Cairo
context causes paint to "jump" when scrolling horizontally and not
scrolling to the end of the text on GTK3.16 and above. Checks for GTK
versions >3.9 and <3.14.8 when making updating event.x from Cairo
context. See bug 535124, and bug 465309.

Tested on GTK3.22 with snippet attached to this patch, sample project in
bug 535124, and the open type dialog, package explorer in a child
Eclipse.

Change-Id: Iaf464c6fee1d5ef288346abbd0ab5a60f20f1a95
Signed-off-by: Xi Yan <xixiyan@redhat.com>