Bug 336238 - Shell.setBounds sometimes fails to set the right location
Calling gtk_widget_show for the first time before gtk_widget_resize
assumes the initial default size for any subsequent calls to
gtk_widget_move. This causes issue with shell.setBounds where the
location can only be moved to a limited bound on the screen
corresponding to the default size by the window manager, even when the
new resized shell is smaller and is able to move to further x, y
position without being cutoff by the screen. The fix is to set the
default size to the smallest possible (1, 1) before gtk_widget_show.

Tested with attached JUnit test.

Change-Id: I3f20a6aaed5972bfda3775816aab96202bae7e5d
Signed-off-by: Xi Yan <xixiyan@redhat.com>