Bug 530138 - [Wayland] Completion details are far away from the completion list
Wayland does not have support for absolute coordinates. When a
GTK_WINDOW_POPUP is attached to another GTK_WINDOW_POPUP parent, it gets
positioned relatively to its parent, causing it to be far away from the
screen when its position in absolute coordinates. The fix is to use
gtk_window_set_transient_for to attach the GTK_WINDOW_POPUP to a
GTK_WINDOW_TOPLEVEL parent already mapped on screen.

Tested with completion list in child Eclipse on Wayland, GTK3.22.

Change-Id: Ic9d0b4f785a12c6a659546fc362fcc8c9b0ddc4c
Signed-off-by: Xi Yan <xixiyan@redhat.com>