Bug 530577 - [Wayland] Popup menus sometimes emit GDK warnings
1) On Wayland, clicking Show In button in Open Resource dialog causes
menu.setVisible to be called twice, once due to MouseDown and once due
to ButtonClicked. This causes two duplicate menus (with different
handles) to popup at the same time. The second one does not have a
top-most parent which emits a GDK warning, and the menu cannot be
closed. The same issue also happens on both X11 and Wayland with toggle
buttons. See bug 528998. The fix is to avoid adding two identical popup
menus in the
queue.

2) Implemented Menu#setLocation() on Wayland by passing the window of
the relative parent shell of menu to gtk_menu_popup_at_rect to position
the menu. Used hasLocation flag for Wayland (see Bug 530204).

Tested on Wayland and X11 with GTK3.22. Clicking "Show In" button in
Open Resource Dialog, and the "Run As..." triangle now opens popup menu
under the button rather than by the mouse. Right click on perspective
button and clicking anywhere now able to close the context menu, whereas
before requires two clicks.

Change-Id: I89110e5d095ca896fa0b0d775e617889ea75912d
Signed-off-by: Xi Yan <xixiyan@redhat.com>